Fix Torch deprecation and ignore Gym warnings

+++ main.py
@@ -115,7 +115,7 @@ if args.models is not '' and os.path.exists(args.models):
   optimiser.load_state_dict(model_dicts['optimiser'])
 planner = MPCPlanner(env.action_size, args.planning_horizon, args.optimisation_iters, args.candidates, args.top_candidates, transition_model, reward_model)
 global_prior = Normal(torch.zeros(args.batch_size, args.state_size, device=args.device), torch.ones(args.batch_size, args.state_size, device=args.device))  # Global prior N(0, I)
-free_nats = torch.full((1, ), args.free_nats, device=args.device)  # Allowed deviation in KL divergence
+free_nats = torch.full((1, ), args.free_nats, dtype=torch.float32, device=args.device)  # Allowed deviation in KL divergence
 
 
 def update_belief_and_act(args, env, planner, transition_model, encoder, belief, posterior_state, action, observation, explore=False):
